Fix pipeline retrieving an item from the ItemSource prematurely
This caused prioritisation not to work properly when a URL of higher priority
was discovered.

This commit changes function signatures in wpull.pipeline.pipeline. In particular,
ItemQueue.put_item has been replaced by put_item_coro, which now takes a coroutine
function as the argument rather than an item, and ItemQueue.get returns that
coroutine function. In addition, put_item_coro also takes a second argument which
can be used as a callback for the producer (relevant mainly for sending the item
back to the producer and for stopping the producer).

If you used any of the wpull.pipeline.pipeline.{ItemQueue,Worker,Producer} classes
directly, you may have to change your code. The Pipeline, however, remains unchanged.