Work around the change in tornadoweb/tornado@84bb2e28
According to that commit, there is a 200 ms delay on macOS when connecting to
localhost and only IPv4 is bound (which is what Tornado does). For that reason,
they replaced "localhost" with "127.0.0.1" in tornado.testing.AsyncHTTPTestCase.get_url.

Because wpull treats "localhost" and "127.0.0.1" as different hosts, this breaks
a variety of tests for different reasons (URLs getting skipped, cookies not set,
etc.) when using Tornado 4.5.3.

As a workaround, this commit restores the previous behaviour of using "localhost".
This may incur a performance penalty during testing on some platforms.